+++ Jerry/Jerry/Coordinates.h
@@ -2,10 +2,10 @@
 class Coordinates
 {
 public:
-	Coordinates(float x, float y);
+	Coordinates(double x, double y);
 	Coordinates();
 	~Coordinates();
 
-	float X = 0;
-	float Y = 0;
+	double X = 0;
+	double Y = 0;
 };
