+++ Jerry/Jerry/Physics.h
@@ -20,13 +20,13 @@ private:
 	static bool WillCollide(Entity* entity, list<WorldBlock>* world, list<Entity>* entities);
 	static bool AreInRange(Coordinates* entOff, Coordinates* entACo, Coordinates* entBCo, Coordinates* xA, Coordinates* xB);
 	static bool AreColliding(Coordinates* a1, Coordinates* a2, Coordinates* b1, Coordinates* b2);
-	static Coordinates* GetCollisionOffset(Entity* entity, Coordinates* xA, Coordinates* xB, float xStep, float yStep, float* minX, float* minY, float* maxX, float* maxY, bool* minXIsEnt, bool* minYIsEnt);
+	static Coordinates* GetCollisionOffset(Entity* entity, Coordinates* xA, Coordinates* xB, double xStep, double yStep, double* minX, double* minY, double* maxX, double* maxY, bool* minXIsEnt, bool* minYIsEnt);
 public:
 	static VelocityVector* OffsetToVector(Coordinates* offset);
-	static VelocityVector* OffsetToVector(float x, float y);
+	static VelocityVector* OffsetToVector(double x, double y);
 	static Coordinates* VectorToOffset(VelocityVector* vector);
-	static Coordinates* VectorToOffset(float velocity, float angle);
-	static float OffsetToAngle(float x, float y);
+	static Coordinates* VectorToOffset(double velocity, double angle);
+	static double OffsetToAngle(double x, double y);
 	
 	static void ApplyPhysics(list<Entity>* ent, list<WorldBlock>* world);
 };
